requirements.txt
python-dotenv
discord.py
notion-client
google-genai
trafilatura